// packages/sdk/src/views/models/timelineTypes.ts
export enum EventStatus {
    /** The event was not sent and will no longer be retried. */
    NOT_SENT = 'not_sent',
    /** The message is being encrypted */
    ENCRYPTING = 'encrypting',
    /** The event is in the process of being sent. */
    SENDING = 'sending',
    /** The event is in a queue waiting to be sent. */
    QUEUED = 'queued',
    /** The event has been sent to the server, but we have not yet received the echo. */
    SENT = 'sent',
    /** The event was cancelled before it was successfully sent. */
    CANCELLED = 'cancelled',
    /** We received this event */
    RECEIVED = 'received',
}

export interface TimelineEvent {
    eventId: string
    localEventId?: string // if this event was created locally and appended before addEvent, this will be set
    eventNum: bigint
    latestEventId: string // if a message was edited or deleted, this will be set to the latest event id
    latestEventNum: bigint // if a message was edited or deleted, this will be set to the latest event id
    status: EventStatus
    createdAtEpochMs: number // created at times are generated client side, do not trust them
    updatedAtEpochMs?: number // updated at times are generated client side, do not trust them
    content: TimelineEvent_OneOf | undefined // TODO: would be great to have this non optional
    fallbackContent: string
    isEncrypting: boolean // local only, isLocalPending should also be true
    isLocalPending: boolean /// true if we're waiting for the event to get sent back from the server
    isSendFailed: boolean
    confirmedEventNum?: bigint
    confirmedInBlockNum?: bigint
    confirmedAtEpochMs?: number // time miniblock containing this event was created, should be around 2 seconds after the event was created
    threadParentId?: string
    replyParentId?: string
    reactionParentId?: string
    isMentioned: boolean
    isRedacted: boolean
    sender: {
        id: string
    }
    sessionId?: string
}

/// a timeline event should have one or none of the following fields set
export type TimelineEvent_OneOf =
    | ChannelCreateEvent
    | ChannelMessageEncryptedEvent
    | ChannelMessageEncryptedRefEvent
    | ChannelMessageEvent
    | ChannelMessageMissingEvent
    | ChannelPropertiesEvent
    | EncryptedChannelPropertiesEvent
    | FulfillmentEvent
    | InceptionEvent
    | InteractionRequestEvent
    | InteractionRequestEncryptedEvent
    | InteractionResponseEvent
    | KeySolicitationEvent
    | MiniblockHeaderEvent
    | MemberBlockchainTransactionEvent
    | PinEvent
    | ReactionEvent
    | RedactedEvent
    | RedactionActionEvent
    | StreamEncryptionAlgorithmEvent
    | StreamMembershipEvent
    | SpaceImageEvent
    | SpaceUpdateAutojoinEvent
    | SpaceUpdateHideUserJoinLeavesEvent
    | TipEvent
    | TokenTransferEvent
    | SpaceReviewEvent
    | UserBlockchainTransactionEvent
    | UserReceivedBlockchainTransactionEvent
    | UnpinEvent
    | ConversationSeedEvent
    | ConversationSeedResponseEvent
    | SetAppAddressEvent
    | UnsetAppAddressEvent

export type TimelineEventWithContent<T extends TimelineEvent_OneOf> = Omit<
    TimelineEvent,
    'content'
> & {
    content: T
}

export enum RiverTimelineEvent {
    ChannelCreate = 'm.channel.create',
    ChannelMessage = 'm.channel.message',
    ChannelMessageEncrypted = 'm.channel.encrypted',
    ChannelMessageEncryptedWithRef = 'm.channel.encrypted_with_ref',
    ChannelMessageMissing = 'm.channel.missing',
    ChannelProperties = 'm.channel.properties',
    EncryptedChannelProperties = 'm.channel.encrypted_properties',
    Fulfillment = 'm.fulfillment',
    Inception = 'm.inception', // TODO: would be great to name this after space / channel name
    InteractionRequest = 'm.interaction_request',
    InteractionRequestEncrypted = 'm.interaction_request.encrypted',
    InteractionResponse = 'm.interaction_response',
    KeySolicitation = 'm.key_solicitation',
    MemberBlockchainTransaction = 'm.member_blockchain_transaction',
    MiniblockHeader = 'm.miniblockheader',
    Pin = 'm.pin',
    Reaction = 'm.reaction',
    RedactedEvent = 'm.redacted_event',
    RedactionActionEvent = 'm.redaction_action',
    SpaceUpdateAutojoin = 'm.space.update_autojoin',
    SpaceUpdateHideUserJoinLeaves = 'm.space.update_channel_hide_user_join_leaves',
    SpaceImage = 'm.space.image',
    SpaceReview = 'm.space.review',
    StreamEncryptionAlgorithm = 'm.stream_encryption_algorithm',
    StreamMembership = 'm.stream_membership',
    TipEvent = 'm.tip_event',
    TokenTransfer = 'm.token_transfer',
    Unpin = 'm.unpin',
    UserBlockchainTransaction = 'm.user_blockchain_transaction',
    UserReceivedBlockchainTransaction = 'm.user_received_blockchain_transaction',
    ConversationSeed = 'm.conversation_seed',
    ConversationSeedResponse = 'm.conversation_seed_response',
    SetAppAddress = 'm.set_app_address',
    UnsetAppAddress = 'm.unset_app_address',
}

export interface MiniblockHeaderEvent {
    kind: RiverTimelineEvent.MiniblockHeader
    miniblockNum: bigint
    hasSnapshot: boolean
}

export interface FulfillmentEvent {
    kind: RiverTimelineEvent.Fulfillment
    sessionIds: string[]
    deviceKey: string
    to: string
    from: string
}

export interface KeySolicitationEvent {
    kind: RiverTimelineEvent.KeySolicitation
    sessionIds: string[]
    deviceKey: string
    isNewDevice: boolean
}

export interface InceptionEvent {
    kind: RiverTimelineEvent.Inception
    creatorId: string
    type?: PayloadCaseType
    spaceId?: string // valid on casablanca channel streams
    appAddress?: string
}

export interface InteractionRequestEvent {
    kind: RiverTimelineEvent.InteractionRequest
    payload?: PlainMessage<InteractionRequestPayload>
    recipient?: string
    threadParentId?: string
}

export interface InteractionRequestEncryptedEvent {
    kind: RiverTimelineEvent.InteractionRequestEncrypted
    error?: DecryptionSessionError
    recipient?: string
    threadParentId?: string
}

export interface InteractionResponseEvent {
    kind: RiverTimelineEvent.InteractionResponse
    response: PlainMessage<InteractionResponse>
}

export interface ChannelCreateEvent {
    kind: RiverTimelineEvent.ChannelCreate
    creatorId: string
    channelId: string
    channelOp?: ChannelOp
    channelSettings?: unknown
}

export interface SpaceUpdateAutojoinEvent {
    kind: RiverTimelineEvent.SpaceUpdateAutojoin
    channelId: string
    autojoin: boolean
}

export interface SpaceUpdateHideUserJoinLeavesEvent {
    kind: RiverTimelineEvent.SpaceUpdateHideUserJoinLeaves
    channelId: string
    hideUserJoinLeaves: boolean
}

export interface SpaceImageEvent {
    kind: RiverTimelineEvent.SpaceImage
}

export interface ReactionEvent {
    kind: RiverTimelineEvent.Reaction
    targetEventId: string
    reaction: string
}

export interface PinEvent {
    kind: RiverTimelineEvent.Pin
    userId: string
    pinnedEventId: string
}

export interface UnpinEvent {
    kind: RiverTimelineEvent.Unpin
    userId: string
    unpinnedEventId: string
}

export interface ConversationSeedEvent {
    kind: RiverTimelineEvent.ConversationSeed
    seedId: string
    query: string
    responseText: string
    proposals: PlainMessage<Proposal>[]
    clarifyingQuestions: string[]
    createdAtMs: bigint
    externalRef?: string
    conciergeId: Uint8Array
    invocations: PlainMessage<SpecialistCapabilityInvocation>[]
    proposalsByAgent: Record<string, PlainMessage<AgentProposalsResponse>>
    selectedProposalIds: string[]
    context?: string
    source?: string
    statusUpdates: string[]
    replyId?: string
}

export interface ConversationSeedResponseEvent {
    kind: RiverTimelineEvent.ConversationSeedResponse
    seedId: string
    seed?: PlainMessage<ProtoConversationSeed>
    response:
        | { case: 'selectedProposal'; value: PlainMessage<Proposal> }
        | { case: 'userMessage'; value: string }
        | { case: undefined; value?: undefined }
    replyId?: string
}

export interface SetAppAddressEvent {
    kind: RiverTimelineEvent.SetAppAddress
    userId: string
    appAddress: string
}

export interface UnsetAppAddressEvent {
    kind: RiverTimelineEvent.UnsetAppAddress
    userId: string
}

export interface StreamEncryptionAlgorithmEvent {
    kind: RiverTimelineEvent.StreamEncryptionAlgorithm
    algorithm?: string
}

export interface ChannelMessageEncryptedEvent {
    kind: RiverTimelineEvent.ChannelMessageEncrypted
    error?: DecryptionSessionError
    appClientAddress?: string
}

export interface ChannelMessageEncryptedRefEvent {
    kind: RiverTimelineEvent.ChannelMessageEncryptedWithRef
    refEventId: string
    appClientAddress?: string
}

export interface ChannelPropertiesEvent {
    kind: RiverTimelineEvent.ChannelProperties
    properties: ChannelProperties
}

export interface EncryptedChannelPropertiesEvent {
    kind: RiverTimelineEvent.EncryptedChannelProperties
    error?: DecryptionSessionError
}

export interface ChannelMessageMissingEvent {
    kind: RiverTimelineEvent.ChannelMessageMissing
    eventId: string
}

// the same as MembershipOp but with a different name
export enum Membership {
    Join = 'join',
    Invite = 'invite',
    Leave = 'leave',
    None = '',
}

export interface StreamMembershipEvent {
    kind: RiverTimelineEvent.StreamMembership
    userId: string
    initiatorId: string
    membership: Membership
    reason?: MembershipReason
    streamId?: string // in a case of an invitation to a channel with a streamId
    appAddress?: string
}

export interface UserBlockchainTransactionEvent {
    kind: RiverTimelineEvent.UserBlockchainTransaction
    transaction: PlainMessage<BlockchainTransaction>
}

export interface UserReceivedBlockchainTransactionEvent {
    kind: RiverTimelineEvent.UserReceivedBlockchainTransaction
    receivedTransaction: PlainMessage<UserPayload_ReceivedBlockchainTransaction>
}

export interface MemberBlockchainTransactionEvent {
    kind: RiverTimelineEvent.MemberBlockchainTransaction
    transaction?: PlainMessage<BlockchainTransaction>
    fromUserId: string
}

export interface TipEvent {
    kind: RiverTimelineEvent.TipEvent
    transaction: PlainMessage<BlockchainTransaction>
    tip: PlainMessage<BlockchainTransaction_Tip>
    transactionHash: string
    fromUserId: string
    refEventId: string
    toUserId: string
}

export interface TokenTransferEvent {
    kind: RiverTimelineEvent.TokenTransfer
    transaction: PlainMessage<BlockchainTransaction>
    transfer: PlainMessage<BlockchainTransaction_TokenTransfer>
    fromUserId: string
    createdAtEpochMs: bigint
    threadParentId: string
}

export interface SpaceReviewEvent {
    kind: RiverTimelineEvent.SpaceReview
    action: BlockchainTransaction_SpaceReview_Action
    rating: number
    comment?: string
    fromUserId: string
}
// ...
export interface ChannelMessageEvent {
    kind: RiverTimelineEvent.ChannelMessage
    threadId?: string
    threadPreview?: string
    replyId?: string
    replyPreview?: string
    body: string
    mentions: OTWMention[]
    editsEventId?: string
    content: ChannelMessageEventContentOneOf
    attachments?: Attachment[]
    appClientAddress?: string
}

export type ChannelMessageEventWithContent<T extends ChannelMessageEventContentOneOf> = Omit<
    ChannelMessageEvent,
    'content'
> & {
    content: T
}

// original event: the event that was redacted
export interface RedactedEvent {
    kind: RiverTimelineEvent.RedactedEvent
    isAdminRedaction: boolean
}

// the event that redacted the original event
export interface RedactionActionEvent {
    kind: RiverTimelineEvent.RedactionActionEvent
    refEventId: string
    adminRedaction: boolean
}